requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=pypi_0
alabaster=0.7.12=py36_0
appdirs=1.4.4=py_0
argh=0.26.2=py36_0
arrow=0.13.1=py36_0
astroid=2.5=py36haa95532_1
astunparse=1.6.3=pypi_0
async_generator=1.10=py36h28b3542_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
autopep8=1.5.6=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py36he774522_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=19.10b0=py_0
bleach=3.3.0=pyhd3eb1b0_0
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2021.4.13=haa95532_1
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py36haa95532_0
cffi=1.14.5=py36hcd4344a_0
chardet=4.0.0=py36haa95532_1003
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cookiecutter=1.7.2=pyhd3eb1b0_0
cryptography=3.4.7=py36h71e12ea_0
cycler=0.10.0=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=py_0
docutils=0.17.1=py36haa95532_1
entrypoints=0.3=py36_0
flake8=3.9.0=pyhd3eb1b0_0
flatbuffers=1.12=pypi_0
future=0.18.2=py36_1
gast=0.4.0=pypi_0
google-auth=1.30.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
imageio=2.9.0=pypi_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=4.0.1=pypi_0
importlib_metadata=3.10.0=hd3eb1b0_0
inflection=0.5.1=py36haa95532_0
install=1.3.4=pypi_0
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.8.0=pyhd3eb1b0_0
jedi=0.17.2=py36haa95532_1
jinja2=2.11.3=pyhd3eb1b0_0
jinja2-time=0.2.0=pyhd3eb1b0_2
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py36haa95532_0
jupyterlab_pygments=0.1.2=py_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
keyring=23.0.1=py36haa95532_0
kiwisolver=1.3.1=pypi_0
lazy-object-proxy=1.6.0=py36h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h6c2663c_0
markdown=3.3.4=pypi_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.3.4=pypi_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36he774522_0
mypy_extensions=0.4.3=py36_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py36_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5.1=pypi_0
numpy=1.19.5=pypi_0
numpydoc=1.1.0=pyhd3eb1b0_1
oauthlib=3.1.0=pypi_0
openssl=1.1.1k=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py36haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
pathspec=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=pypi_0
pip=21.1.1=py36haa95532_0
pluggy=0.13.1=py36haa95532_0
poyo=0.5.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.17.0=pypi_0
psutil=5.8.0=py36h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.6.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.2.0=pyhd3eb1b0_0
pygments=2.9.0=pyhd3eb1b0_0
pylint=2.7.4=py36haa95532_1
pyls-black=0.4.6=hd3eb1b0_0
pyls-spyder=0.3.2=pyhd3eb1b0_0
pynacl=1.4.0=py36h62dcd97_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h6538335_2
pyrsistent=0.17.3=py36he774522_0
pysocks=1.7.1=py36haa95532_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.36.2=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=pypi_0
pywin32=227=py36he774522_1
pywin32-ctypes=0.2.0=py36_1000
pyyaml=5.4.1=py36h2bbff1b_1
pyzmq=20.0.0=py36hd77b12b_1
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.1.10=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.2=pyhd3eb1b0_0
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
regex=2021.4.4=py36h2bbff1b_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rope=0.18.0=py_0
rsa=4.7.2=pypi_0
rtree=0.9.7=py36h2eaa2aa_1
scikit-image=0.17.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=52.0.0=py36haa95532_0
sip=4.19.8=py36h6538335_0
six=1.15.0=py36haa95532_0
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sphinx=4.0.1=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.4=pyhd3eb1b0_0
spyder=5.0.0=py36haa95532_1
spyder-kernels=2.0.1=py36haa95532_0
sqlite=3.35.4=h2bbff1b_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
testpath=0.4.4=pyhd3eb1b0_0
text-unidecode=1.3=py_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tifffile=2020.9.3=pypi_0
tinycss=0.4=pyhd3eb1b0_1002
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py36h2bbff1b_0
traitlets=4.3.3=py36_0
typed-ast=1.4.2=py36h2bbff1b_1
typing_extensions=3.7.4.3=pyha847dfd_0
ujson=4.0.2=py36hd77b12b_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=py36haa95532_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py36_1
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
whichcraft=0.6.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=py36he774522_1
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
